+++ src/main/resources/static/js/services/assignmentsService.js
@@ -1,0 +1,5 @@
+
+export async function fetchAssignments(queryParams = {}){
+    const apiUrl = '/api/assignments';
+    const url = new URL(apiUrl, window.location.origin)
+}
